ai_ats_scanner/requirements.txt
reflex==0.4.8
pymupdf==1.23.25
spacy==3.7.4
scikit-learn==1.4.1.post1
sentence-transformers==2.5.1
ollama==0.1.7
python-dotenv==1.0.1
pandas==2.2.1
python-docx==1.1.0